# docs/user/rest_api/open_api_docu_hanami.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hanami-API-Documentation",
    "contact": {
      "email": "tobias.anker@kitsunemimi.moe"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.9.0"
  },
  "servers": [],
  "paths": {
...
    "/v1alpha/network/{network_uuid}": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Get network",
        "description": "Get information of a network from the database.",
        "parameters": [
          {
            "name": "network_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Uuid",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkResp"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Requested object not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "user_context": []
          }
        ]
      },
      "delete": {
        "tags": [
          "network"
        ],
        "summary": "Delete network",
        "description": "Delete a network from the database.",
        "parameters": [
          {
            "name": "network_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Uuid",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Requested object not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "user_context": []
          }
        ]
      }
    },
...
    "/v1alpha/floating_ip/{floating_ip_uuid}": {
      "get": {
        "tags": [
          "floating_ip"
        ],
        "summary": "Get floating_ip",
        "description": "Get information of a floating_ip from the database.",
        "parameters": [
          {
            "name": "floating_ip_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Uuid",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingIpResp"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Requested object not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "user_context": []
          }
        ]
      },
      "delete": {
        "tags": [
          "floating_ip"
        ],
        "summary": "Delete floating_ip",
        "description": "Delete a floating_ip.\\\n\\\nIf the floating_ip is still attached to a virtual_machine, it is detached first, so the\\\naddress is never released while the torii still translates it.",
        "parameters": [
          {
            "name": "floating_ip_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Uuid",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Requested object not found"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "deprecated": false,
        "security": [
          {
            "user_context": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FloatingIpResp": {
        "title": "FloatingIpResp",
        "type": "object",
        "required": [
          "created_at",
          "created_by",
          "floating_ip",
          "updated_at",
          "updated_by",
          "uuid"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "string"
          },
          "floating_ip": {
            "type": "string",
            "format": "ipv4"
          },
          "internal_ip": {
            "description": "Internal IP-address of the attached virtual_machine. Not set, if the floating IP-address is detached.",
            "type": "string",
            "format": "ipv4",
            "nullable": true
          },
          "network_uuid": {
            "description": "Network of the attached virtual_machine. Not set, if the floating IP-address is detached.",
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "string"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "NetworkResp": {
        "title": "NetworkResp",
        "type": "object",
        "required": [
          "created_at",
          "created_by",
          "name",
          "subnet",
          "updated_at",
          "updated_by",
          "uuid"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "subnet": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_by": {
            "type": "string"
          },
          "uuid": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "user_context": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}